requirements.txt
Flask
Werkzeug
boto3
python-dotenv
gunicorn